renamed checkCircleVsPolygon>checkPolygonVsCircle
It makes more sense with how the algorithm
is built, related to unitVector and alikes.

I'm not sure why there was a checkPolygons
check before returning it; when returning,
the collision check is made another time.
Why isn't this cached ?
In the end i kinda broke checkPolygons, so there
is definitely a purpose : the response is the same
but you can't approach the polygon as close as
earlier. WILL FIX.