fixed Polygon to Circle collisions (bis)
I renamed the collision function so it might be
confusing with previous fix stating i already
fixed that type of collision.

Any ordering of Polygon&Circle collisions are
working now. I added a checkCircleVsPolygon
method, which basically call checkPolygonVsCircle
and fix a few bits on collisionData returned.
This is not the sexiest way of doing it but it
keeps testShapes & collisionData simple.

unitVectors are still broken, WILL FIX.