added drawVector, updated draw_collision_response
It's probably good for debugging to have a
drawVector method.

I updated the sample as i did in my pull. I
wasn't sure of it :
I found the old version of
draw_collision_response a bit verbose and not
displaying enough informations for debugging
but it had the advantage of being explicit
and simple. We can always revert back. I'm now
showing both unitVector & separation on sample1.

separation is in the wrong direction. WILL FIX :)