improved drawing in ShapeDrawer & sample colors
It's a little more verbose now that there is a
startPoint argument in some of the draw methods
but i added that so that openFL could fill a
shape. All the moveTo prevented it.

I also changed a bit the logic in drawVertList.
You were iterating backward probably to save up
a few instructions, i fiddled a bit too much with
it so i kinda broke it but i'll refactor that
later.

I updated colors in usage2 if you don't mind.
Also added a caption in that sample.