usage3 added + drawShape method in ShapeDrawer.hx
drawShape takes any type of Shape and will call
the proper drawing method (Circle/Polygon).

usage3 is a stress test sample. It was more a
collision resolving stress test more than a CPU
one but i obviously didn't resist.

It wasn't an emphasis on the physics part, i know
this is not the point of the library but it makes
debugging easier again.

Collision resolution :
There are a few jitterings but it's good overall
IMO, no implosion, a few slidings but nothing
alarming.

Performances :
I have a fairly powerful CPU and at 25 shapes it
starts to drop below 60fps on flash BUT
this actually means :
25 dynamic * (25 dynamic + 5 static) = 750
collisions per frame.

In the sample i test everything against
everything and that doesn't make much sense in
the real world.

1st improvement would be to resolve collisions by
pair applying separation/2 on shape1 and
-separation/2 on shape2 within one iteration for
collisions between dynamic objects. Need to cache
processed shapes as well. WILL FIX.

2nd improvement and you talked about that in a
recent commit is inactivity detection. Inactive
shapes shouldn't be a collision source (shape1).

And of course in the real word we might do some
pre circle to circle collisions, collision
groups, space partitioning...
So, no worries, the apparent limited performances
are just me being really naive with that sample.
I just felt the need to precise that as i push
it to the repo.

Will work on this later next week.